serverpy/requirements.txt
# Core dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0

# HTTP client
httpx>=0.26.0
requests>=2.31.0  # For sync streaming in threads

# Video/Audio processing
# Note: FFmpeg must be installed on the system

# Python utilities
python-multipart>=0.0.6

# yt-dlp is imported from parent directory
# No need to install separately
